+++ corgos_telegram_bot/modules/settings.py
